refactor: Refatoração de código
