+++ src/participacoes/participacoes.controller.ts
@@ -1,7 +1,7 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put } from '@nestjs/common';
 import { CreateParticipacaoDto } from './dto/create-participacao.dto';
 import { UpdateParticipacaoDto } from './dto/update-participacao.dto';
-import { ParticipacaoService } from './shared/participacao.service/participacao.service';
+import { ParticipacaoService } from './participacao.service';
 
 @Controller('participacoes')
 export class ParcipacoesController {
@@ -13,7 +13,7 @@ export class ParcipacoesController {
   }
 
   @Get(':id')
-  getById(@Param('id') id: string) {
+  getById(@Param('id', new ParseIntPipe()) id: string) {
     return this.participacaoService.getById(Number(id));
   }
 
@@ -23,12 +23,12 @@ export class ParcipacoesController {
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() updateParticipacaoDto: UpdateParticipacaoDto) {
+  update(@Param('id', new ParseIntPipe()) id: string, @Body() updateParticipacaoDto: UpdateParticipacaoDto) {
     return this.participacaoService.update(Number(id), updateParticipacaoDto);
   }
 
   @Delete(':id')
-  delete(@Param('id') id: string) {
+  delete(@Param('id', new ParseIntPipe()) id: string) {
     return this.participacaoService.delete(Number(id));
   }
 }
