+++ src/participacoes/participacoes.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ParcipacoesController } from './participacoes.controller';
-import { ParticipacaoService } from './shared/participacao.service/participacao.service';
+import { ParticipacaoService } from './participacao.service';
 
 @Module({
   controllers: [ParcipacoesController],
